完成基本功能：折线、x/y轴的文字、虚线、加载更多（待优化）

+++ app/src/main/java/wgyscsf/financialcustomerview/fund/FundMode.java
@@ -12,7 +12,7 @@ import wgyscsf.financialcustomerview.utils.RegxUtils;
 public class FundMode {
     //x轴原始时间数据，ms
     public long datetime;
-    public double dataY;
+    public float dataY;
     public String originDataY;
 
     public FundMode(long timestamp, String actual) {
